crud for supplier

+++ database/migrations/2022_09_09_175638_create_order_details_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->unsignedBigInteger('order_id');
             $table->unsignedBigInteger('product_id');
             $table->integer('quantity');
-            $table->float('total_price');
+            $table->integer('total_price');
             $table->timestamps();
              $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
